aviationstack-sam/aviationstack-app/requirements.txt
boto3
pandas
requests
snowflake-connector-python